Convert the charts to a macro
There is a lot of data and we're planning on bringing in more. It's
difficult to manage all different divs across two files. Totally
unreasonable if we can make it a macro and share the contents.

This way, we can modify a single HTML block that will fit all purposes.