Use maths to work out metrics period
We'd like to always try and show ~300 points on the metrics charts,
regardless of the time period that the user has chosen.

AWS CloudWatch is fussy about which time periods you can provide:

https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_MetricStat.html

> the period can be 1, 5, 10, 30, 60, or any multiple of 60

We can use some maths to pick the best one of these values based on the
duration the user asked for and the target of showing 300 datapoints.

The bit that's most confusing here is probably the use of D3's
"bisectLeft" function to pick the best period from an array of allowed
periods. See the (confusing) docs here:

https://github.com/d3/d3-array#bisectLeft

Done as a pair.

Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>