Adjust one year old metrics rule
At the moment, if the user chooses a time period from now and one year
ago, they might find themselves in a position where the minute has
passed, before the request was made and our code will reject it as it's
already 1 year and 1 minute of metrics.

We're adding some headrom in a shape of one week, just to ease the user
experience.