Unify declaration of links
A lot of links across the board don't implement the `govuk-link` class.
Which in theory is not an issue, but means we're not trurly using the
design system and there may be edge cases that Design System covers
where we forget about them. Let's trust our friends from the Design
System team!

Rafal is also going mad with the ":visited" variants of links which
aren't content but more of a navigation function. It appears odd, that
you only ever visit one of say three of the spaces on the list. Adding
the scss will provide us with the eye-candy and potentially better
navigation around the system, and will make Rafal happy!