Remove £ from billing CSVs
The £ character isn't in the ASCII charset, so by including it in the
CSV we're running the risk of other bits of software (e.g. Microsoft
Excel) getting confused about the encoding and showing garbled text.

In practice I'm not sure whether this could actually cause problems -
most text encodings agree that `0xa3` is `£`, and the ones that don't
are mostly from IBM mainframes and other relics of the 1980s.

We have had some reports of weird minus characters n Excel though, and
it seems like there's no real user need for the £ sign if the currency
is in the heading.