Bump nock
Nock now comes with pre-defined types as part of the dependency,
therefore having `@types/nock` is no longer a requirement.

As it turns out, upgrading nock has highlighted an issue in our
codebase, where we're providing not-escaped URL into query parameters.

That means, the URL could end up with two question marks in it, which
older browsers could be confused about...

Adding the escaping functionality to hopefully manage the URL properly
from now on.