Bump openid-client
The new version of a client comes in with some types by default. There
is no need for us to try and maintain our own version of these types.

Some of the API has changed, but the release notes were fairly
reflective of that change.

The real issue came in with the try method in our codebase, that for
tests is hiding the actuall issue, which was that the client now will
validate the token and the IAT and EXP field of the payload need to be
integers instead of floats. Adding some math.round to make sure this is
not an issue for the tests.