Add placeholder null vals to elasticsearch metrics
The charts draw only the metrics you give it, and do not know about the
start/end times.

As such, if we do not add placeholder values, then we will only get
metrics over the time period for data returned which could confuse
users.

Add placeholder values, like in metric-data-getters/cloudwatch so that
the charts will be drawn according to the start/end time rather than the
time period for the metrics returned

> As a user, when I see a metrics chart, I expect to see a metrics chart
for the start/end time, rather than for the metrics returned, so that I
can be sure that my inputted start/end times were taken into account

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>
Co-authored-by: Sebastian <sebastian.schmieschek@digital.cabinet-office.gov.uk>