Reactify the org-users component
This is relatively big commit.

Firstly, it contains additional support for files that can be imported
to our TybeScript files. These are image files.

It contains a new helper, to allow us capitalising strings easily.

And then, it contains a HUGE amount of changes from NJK to TSX, some
involving actual functionality that has been broken... That
functionality is mainly around Permissions, User POST data, validation
and types.

Previously, this was oblivious and quite frankly surprising that it
worked at all...