Fix the instance label
At the moment, the regex sniffing out the instance number is not taking
an effect which is essentially causing the default `001` to kick in.

This is not that great from the user perspective as misleading and
repeated information is being presented.

We're adjusting the Regex allowing us to correctly get the number out of
the label.

As I'm not very good at regex, I've written a test case that will help
establish issues like that in the future...