Fix the instance label
At the moment, the regex sniffing out the instance number is not taking
an effect which is essentially causing the default `001` to kick in.

This is not that great from the user perspective as misleading and
repeated information is being presented.

We're adjusting the Regex allowing us to correctly get the number out of
the label.

As I'm not very good at regex, I've written a test case that will help
establish issues like that in the future...

As it turns out, there are differences between the Redis and
ElasticSearch instances.

ES comes back with an IP address, where as redis has the expected format
containing the ID of the instance, which we'd like to be using.

We're resolving that with a view checking if the string is longer than 3
characters and wrapping it in the abbreviation tag, which hopefully
allows people to access the information if needed.