Add Grid to the Graphs
Personally I find it difficult to look at and follow. I may be wrong,
and I haven't actually discuss this with anyone else. Not certain why
the grid was previously not a thing nor why the left axis line was not
visible.

I'm also adding some headroom to the charts, which will allow the spike
not hit the top border but always stay inside the graph.

- Re-order rendering positions

My OCD when trying to look at the graph from source code is killing me a
little.

By moving those around, we're rendering left to right, top to bottom
pretty much, with an exception for the legend labels.