Fix number formatting issues
There are two issues we observed with the current setup.

1. Large numbers (such as 10000000000000), will go off screen and not
  present itself properly on the chart
2. Some labels often repeat themselves... (10, 20, 30, 40, 50, 50, 50)

During the brief investigation, where the stub data was reproducing the
issue number 2, it came apparent that the d3Format is somewhat faulty.

This came really apparent, when we enabled the default labels for Grid
and have formatted the axis. This caused two values being printed... One
with the incorrect behaviour, and one with correct behaviour.

I believe this is related to the actual format that we have been
passing, for instance I can still reproduce it with `d3Format('.1r')`.

We're creating our own formatters and setting these up in the metrics so
we have control over our own destiny.