Change wording for Graphs
At the moment, there is a helper text on the top of the graphs page,
saying:

> Each point is an average over *2 hours*.

This is vague and sometimes not very well representative of the graphs
themselves.

We are making it a little more descriptive to say:

> Each point on a graph is aggregated over *2 hours* of data.

Which we hope will produce a more meaningful message for the tenants.

Co-authored-by: Claire McNally <claire.mcnally@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne Welch-Richards <toby.lornewelch-richards@digital.cabinet-office.gov.uk>