Increase left axis padding
There is a potential to us to have labels so big, out run the available
space.

Though, we hope it will be `100B` which is 4 characters long, it's
potentially possible to get `999.99GiB` which is quickly 9 characters
long...

Increasing padding, should give some extra room, but will make smaller
labels look ugly.