Fix the menu handle on mobile devices
At the moment, this particular functionality is broken and prevents
administrators using the tool on mobile devices.

We're adding fixes to make sure the menu is mobile-first. This means, we
cannot just extend the govuk-tag class and have to set the values
manually.

We're also making some accessibility changes to the region link, to
indicate more precisely, what it's doing.