Allow administrators to create organisations
We'd like to be deprecating more custom scripts from our repos and move
operational "one time" things to pazmin.

This functionality will allow us to create new organisations that
include the sandbox and tag it with an owner used in different section
of the system.

It will not invite users by default.

Potential improvements could be made around the "Owner" box, as we most
likely have a list of "owners" that we could display instead.

That will however do for now.