Remove any spaces from an email address
It appears that some people copy and paste an email address with a
trailing space or miss type it in the first place.

Whilst our validation sees that as correct email, UAA API would respond
with an error, which we do not handle...

We have the principle of making users pass the first time, therefore
we're stripping out the space instead of presenting an error to them
complaigning about the space.

Co-authored-by: Toby Lorne Welch Richards <toby.lornewelch-richards@digital.cabinet-office.gov.uk>