Remove back-link
Originally, the intention was to move the back-link out of the
main-content element into the template and make it triggerable per page.

However, the back-link should not be used with breadcrumbs, which it
was. I've decided to just deprecate back-links completely, in favor of
breadcrumbs and introduce these where the back-links use to be.