Fix new lines in CSV
The CSV files have only two lines, but a lot of entries in the second
line. This is functionality that use to work fine, before I got to it
and broke it down. Weirdly, I've also modified the tests that were
presumably correct at the time...

Applied fixes to hopefully not let that happen again.