Be nice to Global Auditors accessing user pages
We know [1] that Global Auditors can't see the pages corresponding to individual users. This is part of Cloud Foundry's permission model and not something we really want to change. But until now Global Auditors have got an unfriendly, unclear error message.

This commit changes the page for individual users so that Global Auditors will get a new informative message telling them what to do:

> your "global auditor" permissions do not allow viewing users. sorry. please ask someone who will have more permissions, such as an engineer.

A better fix would be to refine the concept of what a global admin is but that would be a larger project.

[1] https://github.com/alphagov/paas-admin/issues/103