Fix broken test for Global Auditors viewing users
Until now this test has asserted that Global Auditors can see user
pages. The test passed because it mocked all the underlying APIs as
returning HTTP 200 responses, but they actually don't.

The previous commit changed the implementation to return Global
Auditors a prettier error than a Not Found page. This commit updates
the test to match that new implementation.