return "metrics unavailable" page for unknown svcs
the service metrics tab would result in a 500 error if it encountered a
service type that it didn't understand.

there are several ways this could occur:

* A new backing service without metrics implemented yet
* A custom service from a space-scoped broker
* A user provided service

The user-provided-service path was hard coded to display a nicer error
page, but anything else would result in a generic error.

This changes it so that ANY unknown service gets the "metrics
unavailable" page so the user doesn't loose their flow and the tabbed
interface operates as expected.