Use Zendesk Requests API for ticket creation
With Tickets API we couldn't set the user's email address as the
requester, which meant manually updating email for each ticket.

With this change and and the addition of the requester object,
tickets that are created have the user's email address and name.

Additionaly, users now get an email confirmation and email updates.