Introduce quota management
Personally I find it difficult to figure out which region does the
organisation belong to, by first logging in to possibly two pazmins, and
then the CLI, figure out what does the quota offer and set it for the
user before informing them of what has been done to their org.

This should reduce the amount of actions I need to perform each time and
help me move on with my life without distractions.

Probably useless to the rest of the team.