Provide descriptive buttons in calculator
Problem:
The multiple generic '+ Add' and ‘x’ buttons are not descriptive enough
for screen reader users to determine their function or purpose when
navigating out of context.

Also, the aria-label element is spelled wrong which means “Remove” is
not announced to screen reader users; however, multiple generic ‘Remove’
buttons are also not descriptive enough for screen reader users to
determine their function or purpose.

Solution:
- Update 'Add' button for each row with descriptive text to read:
  - for services: "Add selected <serviceName> service plan"
  - for app instances: "Add compute instance with selected configuration"
- Update 'Remove' button for each added row with descriptive text to
read:
  - for services: "Remove <serviceName> <servicePlan> service plan"
  - for app instances: "Remove compute configuration of <number> app
  instances with <selectedMemory> of memory"

Fixes: https://www.pivotaltracker.com/n/projects/1275640/stories/174313713