Update calculator layout
Problem:

The ‘Select options’ column refers to multiple form elements in some
rows which may be difficult for some users to differentiate between;
the columns header does not accurately convey the purpose of the
corresponding data; the header cell spans what appears to be three
columns but has not been marked-up as such.

Solution:
Alternatively, use a definition list similar to the GOV.UK Check answers
Page pattern to present the content on this page.

We have used the https://design-system.service.gov.uk/components/summary-list/
component (definition list) for the updated layout.

Fixes: https://www.pivotaltracker.com/n/projects/1275640/stories/174300123