Introduce API Tokens feature
Users started to come into the support, asking for non-email based users
which could be used for their CI systems to interact with PaaS.

This feature, is simply based on the username and password system, but
relies on specific user naming, in order to defrienciate between the
Users and Tokens.

An Admin and Org Manager should be able to create unlimited number of
these tokens, retreive the Key and Secret, setup the permissions for
Robot User and use the feature as they would with a google group user.