WIP: Add API Token rotation functionality
It would be great if the feature allowed for easy rotation of the
secret.

At the moment, this could be achieved by deleting the user and
re-creating. This however could be problematic when one would want to
replicate the permissions of the deleted user - especially when revoking
the token and not noting what access it used to have.

This feature however, would require either an Admin token, or the Pazmin
tool to have `uaa.admin` permissions, which it currently doesn't have.