Provide a visually-hidden table caption
When screen reader users navigate to the table either on first visit
or when sorting the table,
we need to tell them which column filters are applied.
We do this by providing a table caption that includes month,
any space/service filter and column filter applied and order of sort.

As we only get a date string param available in the Statement function,
we've created a function to convert to month name.