Update metrics page layout
As we're moving to invidiual date and time input for greater
acessibility, which will make the change range container much longer.

This updates the layout so that the graphs and tables take up 2/3 of
the width and the other 1/3 is the range container, for the whole of
the page.

We've also moved the text about the current time period out of the
container to above the metrics.