sqs: add sent/recv baacking service metrics
display charts for number of messages sent/recv for sqs backing services

this is not all the metrics available, it was an initial pass to ensure
that we are able to collect the metrics, and there are no blockers for
quickly iterating this as the needs arrise.

currently only checks the "primary" queue (ignore the "secondary" queue
that is often configured as a dead-letter queue).