Test audit event pages with autoscaler events
We know the app and space audit event pages are erroring when they
include audit events from the autoscaler [1]. This seems to be because
the events are coming back in a way we never anticipated:

    "actor": {
      "type": "user",
      "guid": "app_autoscaler",
      "name": ""
    },

We already don't look up actors whose type is not "user". We expected
that every actor with type "user" would have a GUID that's a UUID, and
probably show up in `paas-accounts`. `app_autoscaler` is a UAA client,
but it's reported with type of "user" too.

This commit adds tests to detect this problem so we can then fix it.

[1] https://github.com/cloudfoundry/app-autoscaler