Use fake AWS credentials when using the stub API
When using the stub API, we have fake responses to AWS API calls, but we still create instances of the client types in code. When this happens, the clients look for AWS credentials in all the normal places.

Sometimes, developers don't have AWS credentials set, or otherwise don't have the `default` profile in place. Since we don't actually NEED the AWS credentials, we should set fakes explicitly when using the stub API. That will stop the AWS clients from searching for credentials.