Fix failing date function tests
When we parse dates that are in the past we round them down differently
depending on how far in the past they are.

In our test for sanised user input we were expecting a non-rounded value.

To fix the failing tests, we've now set all test dates to be far in the
future.