List request as a dependency
At the moment we have a failing build with

```
Error: Cannot find module 'request'
   2021-02-16T14:43:29.32+0000 [APP/PROC/WEB/0] ERR Require stack:
   2021-02-16T14:43:29.32+0000 [APP/PROC/WEB/0] ERR - /home/vcap/app/node_modules/request-promise-native/lib/rp.js
   2021-02-16T14:43:29.32+0000 [APP/PROC/WEB/0] ERR - /home/vcap/app/node_modules/jsdom/lib/jsdom/browser/resources/resource-loader.js
   2021-02-16T14:43:29.32+0000 [APP/PROC/WEB/0] ERR - /home/vcap/app/node_modules/jsdom/lib/api.js
   2021-02-16T14:43:29.32+0000 [APP/PROC/WEB/0] ERR - /home/vcap/app/dist/main.js
```

This is due to a combination of factors, mainly upgrading
`notify-node-client ` which used to use this and have since moved away
and the fact that `jsdom` depend on it though `request-promises-native`
which along with `request` have been deprecated.

To give us some time and have a working deployment until we either
replace jsdom or jsdom [move away from deprecated packages](https://github.com/jsdom/jsdom/issues/3030),
we list request as a dependecy.