Present data in tables
In addition to graphs we're showing data in tables for a11y.

While graph data is plotted directly for source in weekly increments,
table data is shown as max number of <metric> per month for previous
months, excluding current month as we already say "X currently <metric>"
- and also latest value might not be the max number for the month, so
there would be a mismatch.