Add functions to manipulate data
Data comes in weekly increments and that would have been a very long
table for the entire year, so we've decided to show max numbers for
each month.

To do that we have the `exportMaxPerMonthDataValues` function,
that takes the source data, sorts it by descending value column,
and then removes duplicate montly values, ensuring highest montly value
remains.
We could have used Lodash but native javascript works well.

For the organisation data we needed a different data structure
as we're showing the data i ntwo columns per row.

`combineMetrics` function takes in two arrays, utilises the
`exportMaxPerMonthDataValues` to extract max montly values
and finally combines data into a new structure.

To work with dates we have a `formatDate` function, as MomentJs is in
[maintenance mode only](https://momentjs.com/docs/#/-project-status/)
and we need to move away from it's usage slowly.

We utilise native javascript to take in datetime stamp and output
formatted string, in our case "monthName year"