Update line graph legend logic
On the perf page we want to show a label for billable and trial
organisation graphs.

Graph legend logic was set up to only show the leged
if label ended in '-<number>'.

Amend logic will display a custom label text if provided, otherwise
if the label matches previous logic, it will display that.