Stub data api for /performance page
Page expects data to come from a `metrics.json` file. In staging and
prod this is served by the  paas-admin-data` internal app, but for
local development there was none.

This introduces an endpoint as part of the `stub-apis-env` script on
`npm run start:stub-api` command.

Data is pseudo-randomly generated for the last year in weekly
increments, much like the real data.