Add event tracking functions
To measure engagement for users that have accepted cookies we want to
record certain events with either specific labels for targeted elements
or more generic for larger groups .

This script introduces functions to send events for:

- targeted elements with data attributes
- external links to non *.cloud.service subdomains and email links
- error pages (403, 404, 500)

For targeted elements values for the event are read from the data
attributes.

For external links, values are partially set and partially read from
the target element.

For error pages, values are set.