Remove elements from print
GOV.UK Front provides a utility class `govuk-!-display-none-print` to
remove those elements from print.

Adding those classes to elements makes it clear that they are not meant
to be visible in the print output.

If we later deem that we want to additionally prettify print output, we
can remove those and move overrides into the @print block in the main
stylesheet.