Resolve missing css stylesheet
- update css extract plugin config with filename
- update css import and reference in the template file

Mini-css-extract-plugin generates a file and wan'ts to insert the link
to the assets in the browser.
This being a server-side rendered app, it cannot do that, so we need to
manually link to the file.

While webpack can gerate the file with a cache-busting string there is
no reliable way to export that and use in the template.
We have tried with webpack-manifest plugin, but ran into a
circular-dependency contstant rebuild and a test error for an empty file.

While ideally we'd have cache-busting, but our server is already set
without any cache headers (suspect because of client js), so the broser
will request the CSS file everytime, meaning we're not loosing much if
anything.
Happy to accept this compromise until we can spend more time looking at
the whole pipeline.