Fix authentication table markup on team members page
Previously, users with "password" authentication got a visually hidden
"Authentication method:" span and a confusing link (to the same place as
the "Email address" cell one column to the left). Users with "Google" or
"Admin-google" don't get the link, and also don't get the visually
hidden span (which is probably an a11y issue).

Users where `props.privileged` was falsey got a table with an empty
column heading ("" instead of "Authentication") containing a bunch of
<NoTick />s (which are visually hidden, but readable by screen readers).
This changes the logic to skip the entire column (both the heading and
the cells) if the user isn't privileged enough to see the
Authentication.