Filter out only a valid date series
There is a possibility that the series will poses some invalid dates. We
don't want to be dealing with that, as it may produce some bad outcome
for the end user, where for instance, a lot of events happened on the
Thursday, 1 January 1970 00:00:00, and then few or none events on
current timing people actually care about.

Arguably, filtering them out could leave the graph empty, but at this
point I feel like there other issues we have on hands as oppose to
irrational filtering out?