Add additional models to node-zendesk types
Zendesk API allows for [setting email CCs](https://developer.zendesk.com/documentation/ticketing/managing-tickets/creating-and-updating-tickets/#setting-email-ccs)
and [specifying a requester](https://developer.zendesk.com/documentation/ticketing/managing-tickets/creating-and-updating-tickets/#creating-a-ticket-with-a-new-requester)
for a new ticket.

Library node-zendesk that we use doesn't by default have all of those
options, but as it is a thin wrapper for the API, it passes all fields
through.