Include JavaScript for skip link
to improve screen reader announcements

JavaScript for the skip link component to set focus to the linked element, for example, the main content on the page. This helps screen readers read the linked content when users use the skip link component.

add a data-module="govuk-skip-link" attribute to the component HTML. For example:

To check the skip link JavaScript works. select the skip link and check the linked element (usually the <main> element) in the browser has a tabindex attribute.

As per https://github.com/alphagov/govuk-frontend/releases/tag/v4.0.0