Footer markup chanhes
Fixes to the alignment of columns within the footer component, so they now align with our grid. We've also removed the logic that assumes a 2-section layout displays as a 'two-thirds and one-third' layout. Footer sections now display as full-width by default.

check your footer displays as expected and use the standard Design System grid classes on the govuk-footer__section element to set the width.