Replace custom stylelint
Replace custom styling rules with GDS Stylelint Config

In https://github.com/alphagov/paas-admin/pull/2226 we removed
GDS Stylelint Config and replaced it with custom rules, so we would not
be blocked upgrading to v14 (security fixes)

Now, we can move back to it as it has been upgraded to be compatible
with stylelint 14.

Additionally added back two specfic rules for us.