Add ability to fetch requester details on support form submission
When the support forms are submitted with fetch requester details.

If the requester of our support forms is logged into paas-admin,
we want to include their cloudfoundry details in the ticket as a
private comment, so that we can quickly check that they have
appropriate roles.

If they're not logged in or don't have an account, we also include
that in the comment.

On form submission we check the session for user token and check UAA
if the user exists. If not we, return an empty object.

If they do, we do a lookup if they're members of any organisations and
if so what are their roles.

Once all the information is collated, we first create the zendesk ticket
and then update it with the requester details in the private comment.